[#109] split containers into a container/component setup
- when a component needs to use mapStateToProps/mapDispatchTopProps we create a container for it and add our logic there. When we import the components we then import from our container instead of our component
