+++ client/src/containers/home/index.js
@@ -1,19 +1,7 @@
-import React from 'react'
 import { push } from 'react-router-redux'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-
-const Home = props => {
-  const { changePage, message } = props
-
-  return (
-    <div>
-      <h1>Home</h1>
-      <p>{message}</p>
-      <button onClick={() => changePage()}>Go to about page via redux</button>
-    </div>
-  )
-}
+import Home from '../../components/home'
 
 const mapStateToProps = state => {
   return {
